Fix compile break on 32-bit platforms introduced in 02ebd93002ee9c79a0f5899b4f78203d54f31333

+++ lib/CL/pocl_llvm_api.cc
@@ -451,7 +451,7 @@ int pocl_llvm_build_program(cl_program program,
   clang::PrintPreprocessedAction Preprocess;
   success = CI.ExecuteAction(Preprocess);
   char *PreprocessedOut = nullptr;
-  size_t PreprocessedSize = 0;
+  uint64_t PreprocessedSize = 0;
 
   if (success) {
     pocl_read_file(tempfile, &PreprocessedOut, &PreprocessedSize);
@@ -467,7 +467,7 @@ int pocl_llvm_build_program(cl_program program,
   }
 
   pocl_cache_create_program_cachedir(program, device_i, PreprocessedOut,
-                                     PreprocessedSize, program_bc_path);
+                                     static_cast<size_t>(PreprocessedSize), program_bc_path);
 
   POCL_MEM_FREE(PreprocessedOut);
